rename: Analytics 폴더명 변경
